Moved interfaces to a dedicated interfaces folder for better code organization.
